Finish Blog Post Application

+++ controllers/postsController.js
@@ -1,15 +1,15 @@
 const Post = require('../models/post')
 const Comments = require('../models/comment')
 const { body, validationResult } = require('express-validator')
+var fsP = require('fs/promises')
+var fs = require('fs')
+
 
 exports.posts_get = async (req, res, next) => {
-    const { sort } = req.query
-    const posts = await Post.find().sort(sort ? { [sort]: sort[0] == '-' ? -1 : 1 } : null)
+    const { sort, ...queries } = req.query
+    const posts = await Post.find(queries).sort(sort ? { [sort]: sort[0] == '-' ? -1 : 1 } : null)
 
-    res.json({
-        posts,
-        results: posts.length
-    })
+    res.json(posts)
 }
 
 exports.post_post = [
@@ -24,17 +24,19 @@ exports.post_post = [
         .escape()
         .isLength({ min: 1 })
         .withMessage('Invalid title length')
-        .isAlphanumeric()
-        .withMessage('Title contains non-alphanumeric characters')
+    ,
+    body('cover')
+        .optional({ checkFalsy: true })
     , async (req, res, next) => {
         const errors = validationResult(req)
         if (!errors.isEmpty()) {
-            res.sendStatus(400)
+            return res.sendStatus(400)
         }
         try {
             const post = new Post({
                 content: req.body.content,
-                title: req.body.title
+                title: req.body.title,
+                cover: req.file ? req.file.filename : ''
             })
 
             await post.save()
@@ -58,28 +60,43 @@ exports.post_get = async (req, res, next) => {
 
 exports.post_put = [
     body('content')
+        .optional({ checkFalsy: true })
         .trim()
         .escape()
         .isLength({ min: 1 })
         .withMessage('Content Invalid Length')
     ,
     body('title')
+        .optional({ checkFalsy: true })
         .trim()
         .escape()
         .isLength({ min: 1 })
         .withMessage('Invalid title length')
-        .isAlphanumeric()
-        .withMessage('Title contains non-alphanumeric characters')
+    ,
+    body('isPublished')
+        .optional({ checkFalsy: true })
+        .isBoolean()
     , async (req, res, next) => {
         const errors = validationResult(req)
         if (!errors.isEmpty()) {
-            res.sendStatus(400)
+            return res.status(400).send(errors)
         }
         try {
+            console.log(req.body)
+            const oldPost = await Post.findById(req.params.postid)
+            if (req.body.cover == 'none' && oldPost.cover !== '') {
+                const filepath = oldPost.cover
+                const path = './uploads/' + filepath
+                if (fs.existsSync(path)) {
+                    await fsP.unlink(path)
+                }
+            }
             const updatedPost = new Post({
                 _id: req.params.postid,
-                content: req.body.content,
-                title: req.body.title
+                content: req.body.content ? req.body.content : oldPost.content,
+                title: req.body.title ? req.body.title : oldPost.title,
+                cover: req.body.cover == 'none' ? '' : req.file ? req.file.filename : oldPost.cover,
+                isPublished: typeof req.body.isPublished == 'boolean' ? req.body.isPublished : oldPost.isPublished
             })
 
             const post = await Post.findByIdAndUpdate(req.params.postid, updatedPost, { new: true })
@@ -92,8 +109,16 @@ exports.post_put = [
 
 exports.post_delete = async (req, res, next) => {
     try {
+
         await Comments.deleteMany({ post: req.params.postid })
         const removedPost = await Post.findByIdAndRemove(req.params.postid)
+        const filepath = removedPost.cover
+        if (filepath !== '') {
+            const path = './uploads/' + filepath
+            if (fs.existsSync(path)) {
+                await fsP.unlink(path)
+            }
+        }
         res.json(removedPost)
     } catch (error) {
         next(error)
@@ -102,11 +127,8 @@ exports.post_delete = async (req, res, next) => {
 
 exports.comments_get = async (req, res, next) => {
     try {
-        const comments = await Comments.find({ post: req.params.postid })
-        res.json({
-            comments,
-            results: comments.length
-        })
+        const comments = await Comments.find({ post: req.params.postid }).populate('author').sort({ create_date: -1 })
+        res.json(comments)
     } catch (error) {
         next(error)
     }
@@ -122,7 +144,7 @@ exports.comment_post = [
         const errors = validationResult(req)
 
         if (!errors.isEmpty()) {
-            res.sendStatus(400)
+            return res.sendStatus(400)
         }
 
         try {
@@ -132,13 +154,14 @@ exports.comment_post = [
                 res.status(404).send('The requested post does not exist')
             }
 
-            const comment = new Comment({
+            const comment = new Comments({
+                author: req.user._id,
+                post: req.params.postid,
                 content: req.body.content
             })
 
             await comment.save()
-            
-            res.json(comment)
+            return res.json(await comment.populate('author'))
         } catch (error) {
             next(error)
         }
@@ -146,6 +169,10 @@ exports.comment_post = [
 ]
 
 exports.comment_delete = async (req, res, next) => {
+    const comment = await Comments.findById(req.params.id)
+    if (!comment.author.equals(req.user._id) && !req.user.isAdmin) {
+        return res.sendStatus(403)
+    }
     try {
         const removedComment = await Comments.findByIdAndRemove(req.params.postid)
         res.json(removedComment)
@@ -161,4 +188,12 @@ exports.comments_delete = async (req, res, next) => {
     } catch (error) {
         next(error)
     }
+}
+
+exports.post_publish = async (req, res, next) => {
+
+}
+
+exports.post_unpublish = async (req, res, next) => {
+
 }
